:recycle: refactor!: reorganize the release notes structure
