+++ .github/workflows/create-and-publish.yml
@@ -223,7 +223,7 @@ jobs:
       # Generate release notes for GitHub Release
       # --------------------------------------------------------
       - name: Generate GitHub Release Notes
-        run: node bin/generate-release-notes.js
+        run: node bin/release-notes.js generate
 
       # --------------------------------------------------------
       # Create GitHub Release with notes and attach built assets
