+++ lib/dry-run.js
@@ -3,7 +3,7 @@ import path from "node:path";
 
 import { generateChangelog } from "./changelog/generate.js";
 import { config } from "./config.js";
-import { generateReleaseNotes } from "./release-notes.js";
+import { generateReleaseNotes } from "./release-notes/generate.js";
 import { computeVersion } from "./version/compute.js";
 
 /* ===========================
