+++ lib/release-notes/generate.js
@@ -1,15 +1,15 @@
 import fs from "node:fs";
 import path from "node:path";
 
-import { config } from "./config.js";
-import { getLastTag } from "./git.js";
+import { config } from "../config.js";
+import { getLastTag } from "../git.js";
 import {
   ensureGhCLI,
   getRepoFromGit,
   getDefaultBranch,
   generateReleaseNotesViaAPI,
-} from "./github.js";
-import { computeVersion } from "./version/compute.js";
+} from "../github.js";
+import { computeVersion } from "../version/compute.js";
 
 /* ===========================
  * Core API (Programmatic)
